Implemented early version of session_state save as pickle
